+++ bugbridge/src/main/java/com/cdac/bugbridge/dto/UserResponse.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class UserResponse {
 
-  private Integer id;
+  private Long id;
   private String name;
   private String email;
   private String role;
@@ -22,7 +22,7 @@ public class UserResponse {
   }
 
   // Constructor
-  public UserResponse(Integer id, String name, String email, String role) {
+  public UserResponse(Long id, String name, String email, String role) {
     this.id = id;
     this.name = name;
     this.email = email;
@@ -30,11 +30,11 @@ public class UserResponse {
   }
 
   // Getters and Setters
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
