+++ bugbridge/src/main/java/com/cdac/bugbridge/repository/UserRepository.java
@@ -13,20 +13,20 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     // find user by email
     Optional<User> findByEmail(String email);
 
     @Modifying
     @Query("UPDATE User u SET u.name = :name, u.email = :newEmail, u.role = :role WHERE u.id = :userId")
-    int updateById(@Param("userId") Integer uniqueId, @Param("newEmail") String newEmail,
+    int updateById(@Param("userId") Long uniqueId, @Param("newEmail") String newEmail,
             @Param("name") String name, @Param("role") UserRole role);
 
     // deleting user by ID
     @Modifying
     @Query("DELETE from User u WHERE u.id = :userId")
-    int deleteUserById(@Param("userId") Integer userId);
+    int deleteUserById(@Param("userId") Long userId);
 
     // finding users by role
     List<User> findByRole(UserRole role);
